pyproject.toml
[project]
name = "mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.10",
    "fastmcp>=2.3.0rc1",
    "google-adk>=0.0.1",
]
